split code to files
