+++ hw09_struct_validator/validator.go
@@ -3,8 +3,6 @@ package hw09_struct_validator //nolint:golint,stylecheck
 import (
 	"fmt"
 	"reflect"
-	"regexp"
-	"strconv"
 	"strings"
 )
 
@@ -150,269 +148,4 @@ func validateFieldSlice(errs ValidationErrors, typ reflect.StructField, value re
 type fieldRules interface {
 	validate(errs ValidationErrors, value reflect.Value) ValidationErrors
 	validateSlice(errs ValidationErrors, value reflect.Value) ValidationErrors
-}
-
-// string validations
-
-func fillStringRules(field, value string) (fieldRules, error) {
-	rules := &stringRules{
-		field: field,
-	}
-	strs := strings.Split(value, "|")
-	for _, str := range strs {
-		pair := strings.Split(str, ":")
-		if len(pair) != 2 {
-			return nil, ErrIncorrectUse{reason: IncorrectCondition, field: field, rule: str}
-		}
-
-		ruleName := pair[0]
-		ruleValue := pair[1]
-		var rule stringRule
-		var err error
-		switch ruleName {
-		case "len":
-			rule, err = newStrLen(ruleValue)
-		case "regexp":
-			rule, err = newStrRegexp(ruleValue)
-		case "in":
-			rule = newStrIn(ruleValue)
-		default:
-			return nil, ErrIncorrectUse{reason: UnknownRule, field: field, rule: ruleName}
-		}
-		if err != nil {
-			return nil, ErrIncorrectUse{reason: IncorrectCondition, field: field, rule: ruleName, err: err}
-		}
-		rules.rules = append(rules.rules, rule)
-	}
-	return rules, nil
-}
-
-type stringRule interface {
-	validate(value string) error
-}
-
-type stringRules struct {
-	field string
-	rules []stringRule
-}
-
-func (r *stringRules) validate(errs ValidationErrors, value reflect.Value) ValidationErrors {
-	val := value.String()
-	for _, rule := range r.rules {
-		err := rule.validate(val)
-		if err != nil {
-			errs = append(errs, ValidationError{Field: r.field, Err: err})
-		}
-	}
-	return errs
-}
-
-func (r *stringRules) validateSlice(errs ValidationErrors, value reflect.Value) ValidationErrors {
-	for i := 0; i < value.Len(); i++ {
-		errs = r.validate(errs, value.Index(i))
-	}
-	return errs
-}
-
-type strLen struct {
-	cond int
-}
-
-func newStrLen(value string) (*strLen, error) {
-	val, err := strconv.ParseInt(value, 10, 0)
-	if err != nil {
-		return nil, err
-	}
-	return &strLen{cond: int(val)}, nil
-}
-
-func (s strLen) validate(value string) error {
-	if len(value) == s.cond {
-		return nil
-	}
-	return fmt.Errorf("string length %d is not equal to required %d", len(value), s.cond)
-}
-
-type strRegexp struct {
-	cond *regexp.Regexp
-}
-
-func newStrRegexp(value string) (*strRegexp, error) {
-	rg, err := regexp.Compile(value)
-	if err != nil {
-		return nil, err
-	}
-	return &strRegexp{cond: rg}, nil
-}
-
-func (s strRegexp) validate(value string) error {
-	if s.cond.MatchString(value) {
-		return nil
-	}
-	return fmt.Errorf("string `%s` does not match the regexp `%v`", value, s.cond)
-}
-
-type strIn struct {
-	cond []string
-}
-
-func newStrIn(value string) *strIn {
-	val := strings.Split(value, ",")
-	return &strIn{cond: val}
-}
-
-func (s strIn) validate(value string) error {
-	if stringContains(s.cond, value) {
-		return nil
-	}
-	return fmt.Errorf("string `%s` is not included in the specified set %v", value, s.cond)
-}
-
-// int validations
-
-func fillIntRules(field, value string) (fieldRules, error) {
-	rules := &intRules{
-		field: field,
-	}
-	strs := strings.Split(value, "|")
-	for _, str := range strs {
-		pair := strings.Split(str, ":")
-		if len(pair) != 2 {
-			return nil, ErrIncorrectUse{reason: IncorrectCondition, field: field, rule: str}
-		}
-
-		ruleName := pair[0]
-		ruleValue := pair[1]
-		var rule intRule
-		var err error
-		switch ruleName {
-		case "min":
-			rule, err = newIntMin(ruleValue)
-		case "max":
-			rule, err = newIntMax(ruleValue)
-		case "in":
-			rule, err = newIntIn(ruleValue)
-		default:
-			return nil, ErrIncorrectUse{reason: UnknownRule, field: field, rule: ruleName}
-		}
-		if err != nil {
-			return nil, ErrIncorrectUse{reason: IncorrectCondition, field: field, rule: ruleName, err: err}
-		}
-		rules.rules = append(rules.rules, rule)
-	}
-	return rules, nil
-}
-
-type intRule interface {
-	validate(value int64) error
-}
-
-type intRules struct {
-	field string
-	rules []intRule
-}
-
-func (r *intRules) validate(errs ValidationErrors, value reflect.Value) ValidationErrors {
-	val := value.Int()
-	for _, rule := range r.rules {
-		err := rule.validate(val)
-		if err != nil {
-			errs = append(errs, ValidationError{Field: r.field, Err: err})
-		}
-	}
-	return errs
-}
-
-func (r *intRules) validateSlice(errs ValidationErrors, value reflect.Value) ValidationErrors {
-	for i := 0; i < value.Len(); i++ {
-		errs = r.validate(errs, value.Index(i))
-	}
-	return errs
-}
-
-type intMin struct {
-	cond int64
-}
-
-func newIntMin(value string) (*intMin, error) {
-	val, err := strconv.ParseInt(value, 10, 0)
-	if err != nil {
-		return nil, err
-	}
-	return &intMin{cond: val}, nil
-}
-
-func (s intMin) validate(value int64) error {
-	if value >= s.cond {
-		return nil
-	}
-	return fmt.Errorf("number %d is less than specified %d", value, s.cond)
-}
-
-type intMax struct {
-	cond int64
-}
-
-func newIntMax(value string) (*intMax, error) {
-	val, err := strconv.ParseInt(value, 10, 0)
-	if err != nil {
-		return nil, err
-	}
-	return &intMax{cond: val}, nil
-}
-
-func (s intMax) validate(value int64) error {
-	if value <= s.cond {
-		return nil
-	}
-	return fmt.Errorf("number %d is greater than specified %d", value, s.cond)
-}
-
-type intIn struct {
-	cond []int64
-}
-
-func newIntIn(value string) (*intIn, error) {
-	val, err := strsToInts64(strings.Split(value, ","))
-	if err != nil {
-		return nil, err
-	}
-	return &intIn{cond: val}, nil
-}
-
-func (s intIn) validate(value int64) error {
-	if intContains(s.cond, value) {
-		return nil
-	}
-	return fmt.Errorf("number %d is not included in the specified set %v", value, s.cond)
-}
-
-func stringContains(slice []string, str string) bool {
-	for _, v := range slice {
-		if v == str {
-			return true
-		}
-	}
-	return false
-}
-
-func intContains(slice []int64, str int64) bool {
-	for _, v := range slice {
-		if v == str {
-			return true
-		}
-	}
-	return false
-}
-
-func strsToInts64(strs []string) ([]int64, error) {
-	result := make([]int64, 0, len(strs))
-	for _, v := range strs {
-		val, err := strconv.ParseInt(v, 10, 0)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, val)
-	}
-	return result, nil
 }
